pyproject.toml
[tool.poetry]
name = "dscc"
version = "0.1.0"
description = "databricks security content center"
authors = ["Derek King <derekking001@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
dscc = "dscc.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
fire = "^0.7.0"
pyyaml = "^6.0.2"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
fastapi = "^0.115.12"
jinja2 = "^3.1.6"
redis = "^5.2.1"
python-multipart = "^0.0.20"
pandas = "^2.2.3"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
